requirements.txt
tensorflow>=2.12.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0
scikit-learn>=0.24.0
flask>=2.0.0
flask-cors>=3.0.0
werkzeug>=2.0.0
gunicorn>=20.1.0
tqdm==4.66.1
requests==2.31.0
gdown==4.7.1
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0 